Reordered command syntax
