+++ qpc_tools/cli/ansible/install/roles/cli/tasks/configure.yml
