+++ qpc_tools/cli/ansible/install/roles/cli/tasks/init_vars.yml
