+++ qpc_tools/cli/ansible/install/roles/cli/tasks/install_offline.yml
