+++ qpc_tools/cli/ansible/install/roles/cli/tasks/load_rpm_online.yml
