+++ qpc_tools/cli/ansible/install/roles/cli/tasks/validate_install.yml
