+++ qpc_tools/cli/ansible/install/roles/init/tasks/init_vars.yml
