+++ qpc_tools/release.py
@@ -5,4 +5,4 @@ AUTHOR_EMAIL = 'qpc@redhat.com'
 PKG_NAME = 'qpc-tools'
 BIN_SCRIPT = 'bin/qpc-tools'
 URL = 'https://github.com/quipucords/qpc-tools'
-PLAYBOOK_PATH = ''
+PYTHON_SRC_PATH = ''
