+++ qpc_tools/server/ansible/install/roles/init/tasks/init_vars.yml
