+++ qpc_tools/server/ansible/install/roles/server/tasks/download_quipucords_image_online.yml
