+++ qpc_tools/server/ansible/install/roles/server/tasks/init_vars_online.yml
