+++ qpc_tools/server/ansible/install/roles/server/tasks/uninstall_docker_containers.yml
