+++ qpc_tools/server/ansible/install/roles/server_docker/tasks/init_vars.yml
