+++ qpc_tools/server/ansible/install/roles/server_docker/tasks/install_offline.yml
