+++ qpc_tools/server/ansible/install/roles/server_docker/tasks/load_docker_rpm_online.yml
