+++ qpc_tools/server/ansible/install/roles/server_docker/tasks/load_postgres_image_offline.yml
