+++ qpc_tools/server/ansible/install/roles/server_docker/tasks/start_quipucords_container.yml
