+++ qpc_tools/server/ansible/install/roles/server_podman/tasks/init_vars.yml
