+++ qpc_tools/server/ansible/install/roles/server_podman/tasks/load_postgres_image_offline.yml
