+++ vm_helper_files/Makefile
@@ -68,7 +68,7 @@ clean:
 	scripts/cleanup.sh
 
 install-local-tools: online-prep
-	sed -i "s?PLAYBOOK_PATH = ''?PLAYBOOK_PATH = '/qpc_tools/qpc_tools/install/ansible/'?g" qpc_tools/release.py
+	sed -i "s?PYTHON_SRC_PATH = ''?PYTHON_SRC_PATH = '/qpc_tools/qpc_tools/'?g" qpc_tools/release.py
 	python3 setup.py build
 	@if grep -q -i "release 8" /etc/redhat-release; then \
 	python3 setup.py install --prefix /usr; \
